resume_screening/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pdfminer.six==20221105
python-docx==1.1.0
nltk==3.9.1
scikit-learn==1.8.0
pandas==3.0.1
werkzeug==3.1.1
numpy==2.4.2
scipy==1.17.1
google-generativeai==0.3.1
python-dotenv==1.0.0
